Decompose controller.go
